Use async/await to introduce a provider concept in addition to a stream concept.
